+++ src/options/index.ts
@@ -2,6 +2,7 @@ import { OptionsInterface } from '../plugin';
 import { isObject } from '../utils';
 import normalizeContentsConverter from './contentsConverter';
 import { normalize as normalizeFilename2urlPath } from './filename2urlPath';
+import normalizeMetadataUpdater from './metadataUpdater';
 import { normalize as normalizePlugins } from './plugins';
 
 export function normalizeOptions(
@@ -21,6 +22,7 @@ export function normalizeOptions(
         ...opts,
         plugins: normalizePlugins(opts.plugins),
         filename2urlPath: normalizeFilename2urlPath(opts.filename2urlPath),
+        metadataUpdater: normalizeMetadataUpdater(opts.metadataUpdater),
         contentsConverter: normalizeContentsConverter(opts.contentsConverter),
     };
 }