+++ test/options/fixtures/async-mod.update-metadata.js
@@ -1,0 +1,11 @@
+const sleep = waitSeconds =>
+  new Promise(resolve => {
+    setTimeout(() => {
+      resolve();
+    }, waitSeconds * 1000);
+  });
+
+module.exports = async (_, filedata) => {
+  await sleep(0.5);
+  filedata.x = 42;
+};