+++ test/options/fixtures/mod.update-metadata.js
@@ -1,0 +1,3 @@
+module.exports = (_, filedata) => {
+  filedata.x = 42;
+};